requirements.txt

#oup_dbt dag
dbt-core==1.5.3
dbt-postgres==1.5.3
dbt-airflow==2.9.1

#1c connector
aio-pika==9.1.4
asyncpg==0.28.0
loguru==0.7.2
pandas==2.0.3
pydantic==2.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.19
